Fix for http://jira.codehaus.org/browse/MGWT-285
This patch reorders the entries in the classpath so that the POM
outputDirectory appears first in the classpath and the java and
resources directories next. Otherwise, the maven resource plugin will
indeed filter the resources as asked and put them in the
outputDirectory, but the filtered files will be shadowed by the
original definition in resources directories.