Update GWT to 2.8.0-SNAPSHOT, remove experimental parameter jsInteropMode and remove closureCompiler from IT
GWT 2.8 do not support closureCompiler flag, so has been removed from test. The jsInteropMode was added in 2.7 but jsInterop has it initial release now with 2.8, so this parameter was experimental and finaly not used, so has been removed (this makes some projects fails after update to 2.8, but I think it is acceptable, just need to remove this param and use generateJsInteropExport if needed.